working on decentralized bot form and edit page
