+++ pages/bots/_id/index.vue
